# apps/docs/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FFDB HTTP API",
    "version": "1.0.0",
    "description": "Versioned management, project SQL, authentication, sync, object-storage, email-template, and operations API. Error responses use a stable code/message/request_id envelope. Mutation routes marked with x-ffdb-idempotency require an Idempotency-Key header."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current FFDB deployment"
    }
  ],
...
  "paths": {
...
    "/v1/projects/{project_id}/transaction": {
      "parameters": [
        {
          "$ref": "#/components/parameters/ProjectId"
        }
      ],
      "post": {
        "tags": [
          "Data"
        ],
        "operationId": "transaction",
        "security": [
          {
            "projectBearer": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonObject"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "developerBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque",
        "description": "Opaque platform developer access token."
      },
      "userBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Project end-user access token."
      },
      "projectBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT or API key",
        "description": "Project end-user JWT or scoped developer API key as permitted by the route."
      }
    },
    "parameters": {
...
      "ProjectId": {
        "name": "project_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        }
      },
...
    },
    "requestBodies": {
      "JsonObject": {
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      }
    },
    "responses": {
      "Success": {
        "description": "Successful JSON response",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
      "Error": {
        "description": "Stable FFDB error envelope",
        "headers": {
          "X-Request-Id": {
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          "Retry-After": {
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "request_id": {
                "type": "string",
                "format": "uuid"
              },
              "details": {
                "type": [
                  "object",
                  "null"
                ],
                "additionalProperties": true
              }
            }
          }
        }
      }
    }
  }
}